rename: Car, CommaSeparatedListBuilder 모델로 이동
