+++ src/main/kotlin/racingcar/controller/FinalWinner.kt
@@ -1,5 +1,7 @@
 package racingcar.controller
 
+import racingcar.model.Car
+
 class FinalWinner {
     fun decideWinner(cars: List<Car>): List<String> {
         val winnersPosition = cars.maxOf { it.position.length }
