+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,4 +1,4 @@
-package racingcar.controller
+package racingcar.model
 
 class Car(val name: String){
     var position: String = ""
